requirements.txt
asgiref==3.8.1
Django==4.1.13
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
django-model-utils~=4.5.0
django-debug-toolbar==4.0.0
